feat: adds full screen toggle
currently, clicking the 'reset' button sets the CSS back to approximate original percentages the components took up in the UI. However, this results in some buggy behavior if you try to resize the screen at this point. For example, the React widget will resize to only a portion of the width of the full UI if you change the viewport width.