fix: hooks up front end fetch requests to back end API.
The requests from the front end plugin were failing to contact the
back end plugin for a number of reasons.

1. The paths used in the fetch API calls, eg: '/api/fileglancer/file-share-paths
were relying on the Fileglancer API being served from the root of the
domain. This is not the case for juypterhub, where the application api
is served from a sub path, eg: /jupyter/user/<username>/api/fileglancer/file-share-paths.
To fix this, the getAPIPathRoot function was added to the useFileBrowser
hook. This dynamically decides which API path to use based on whether
the code is being run on a standalone jupyterlab instance or on
jupyterhub.

2. Once corrected, the API requests were not authorized and returned a
403 error. This is because the cookies were not being sent with the
fetch request. Adding the credentials: 'include' option to the fetch
request helped with that. The next issue was that all the requests
sent via a javascript fetch on jupyterhub require an XSRF token in
the header. To fix this I added the react-cookie component and
extracted the token from the _xsrf cookie and added it to the fetch
headers.

3. The URL was being messed up by the react router code as it assumed
that it was being served at the root of the domain. This was fine on
the jupyter lab server, where it would rewrite the
http://localhost:8888/lab URL to http://localhost:8888/lab/files. On
the jupyterhub server, the user specific part of the URL was being
removed when a user clicked on a link internal to the extension. This
caused
https://fileglancer-dev.int.janelia.org/jupyter/user/clementsj/lab
to become
https://fileglancer-dev.int.janelia.org/lab/files
which subsequently broke the code trying to figure out what the base
URL should be for the API calls. I added some code to determine the
base URL when the application mounts. This is used in the BrowserRouter
to make sure all the relative URLs are appended to that base URL.