refactor: Converts backend extension to an ExtensionApp class
Initialization of settings and handlers has been moved from the __init__
and handlers files into the app.py file to centralize configuration.

Adds the '/fg' url to the server, which will be the future home of the
SPA version of the fileglancer UI.

Adds a StaticHandler to serve up static content. This will be used by the
UI when it is compiled as a stand alone single page application.

Adds a VersionHandler to serve up the extension version number, which is
helpful when debugging during deployment.