feat(favorites component): adds favorites list to sidebar, and initial click functionality
clicking on a zone favorite opens the corresponding zone in the zone list. Clicking on a file share path favorite opens the corresponding zone in the zone list and navigates to the files page, displaying the files found at that file share path. functionality that still needs to be added: directory favorites; hooking up the filter to also filter the favorites