feat: adds functionality to favorite/unfavorite directories
clicking on the directory favorite in the sidebar navigates to the directory in the file list, and opens the corresponding zone in the sidebar zone list. The user uses the context menu to favorite/un-favorite directories. To dos: hooking up the filter to also filter the favorites; fix dark styles for sidebar; fix display styles (e.g., how large, default open/closed) for sidebar widgets)