fix: restores build steps to get fileglancer Icon back in the lab.
The fileglancer icon was not showing up in the lab launcher after the
modifications to get the frontend code served from the backend
extension. This restores the icon, by replicating the process used to
build the frotned extension, but only supplying it with the code to
redirect to the /fg/ url.