refactor: changes currentFileOrFolder to currentFolder
check if the filePath url param results in a directory. If not, remove last segment of the path to hopefully get a directory. We only want to set the currentFolder because the file browser can only navigate within folders - there is nothing to show in the file browser if you navigate to a file