feat: Adds Fathom Analytics and some tracking events
The analytics code has been added to site and will track page views by
default. For events, we need to add the trackEvent call to functions
that need to be tracked. This can be done in the way it is in this
commit, or we could possibly come up with a way to tag onClick events,
so that they trigger the trackEvent call, without adding additional code
to them.