fix: Adds Auto logout when a users session expires
To prevent the site from exhibiting abnormal behavior when a users
session expires, I added checkSessionValidity function. The reason I had
to add this function is because there are some API end points that
return a 403 even when the session is still valid. To prevent logging a
user out from a valid 403, the checkSessionValidity function makes a
request to the /api/fileglancer/profile endpoint, which should always
return a 200 OK response if the users session is valid. If that request
returns a 403 response, then the user is redirected to the logout url
and is then asked to log back in again. Once can test this by modifying
or deleting the jupyterhub-session-id cookie and then navigating to
another page on the site, or just waiting for the notifications up to
fire, which occurs once a minute.