fix: Moves retry countdown to CentralServerDownOverlay
Having the countdown code in the CentralServerHealthContext was causing
the site to re-render every second. This resulted in a lot of calls to
the layout storage code, which generated a lot of logs and network
traffic. Moving the countdown to the CentralServerDownOverlay causes
only that component to re-render and drastically reduces the resources
needed.